+++ src/main/java/gregtech/common/tools/ToolPlow.java
@@ -36,13 +36,13 @@ public class ToolPlow extends GTTool {
     }
 
     @Override
-    public boolean isMinableBlock(Block aBlock, byte aMetaData) {
+    public boolean isMinableBlock(Block aBlock, int aMetaData) {
         return GTToolHarvestHelper.isAppropriateTool(aBlock, aMetaData, "plow")
             || GTToolHarvestHelper.isAppropriateMaterial(aBlock, Material.snow, Material.craftedSnow);
     }
 
     @Override
-    public float getMiningSpeed(Block aBlock, byte aMetaData, float aDefault, EntityPlayer aPlayer, World worldObj,
+    public float getMiningSpeed(Block aBlock, int aMetaData, float aDefault, EntityPlayer aPlayer, World worldObj,
         int aX, int aY, int aZ) {
         // Speed nerf for using AOE tools to break single block
         if (aPlayer != null && aPlayer.isSneaking()) {
@@ -53,7 +53,7 @@ public class ToolPlow extends GTTool {
 
     @Override
     public int convertBlockDrops(List<ItemStack> aDrops, ItemStack aStack, EntityPlayer aPlayer, Block aBlock, int aX,
-        int aY, int aZ, byte aMetaData, int aFortune, boolean aSilkTouch, BlockEvent.HarvestDropsEvent aEvent) {
+        int aY, int aZ, int aMetaData, int aFortune, boolean aSilkTouch, BlockEvent.HarvestDropsEvent aEvent) {
         int rConversions = 0;
         if ((this.sIsHarvestingRightNow.get() == null) && ((aPlayer instanceof EntityPlayerMP))) {
             this.sIsHarvestingRightNow.set(this);
