+++ src/main/java/gregtech/common/tools/ToolSword.java
@@ -84,7 +84,7 @@ public class ToolSword extends GTTool {
     }
 
     @Override
-    public boolean isMinableBlock(Block aBlock, byte aMetaData) {
+    public boolean isMinableBlock(Block aBlock, int aMetaData) {
         return GTToolHarvestHelper.isAppropriateTool(aBlock, aMetaData, "sword")
             || GTToolHarvestHelper.isAppropriateMaterial(
                 aBlock,
