+++ src/main/java/gregtech/common/tools/pocket/ToolPocketMultitool.java
@@ -44,7 +44,7 @@ public class ToolPocketMultitool extends GTTool {
     }
 
     @Override
-    public boolean isMinableBlock(Block aBlock, byte aMetaData) {
+    public boolean isMinableBlock(Block aBlock, int aMetaData) {
         return false;
     }
 }
